self.ne might be a float
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@302 6419b465-d902-0410-997c-be2266baf723

+++ gridpaw/occupations.py
@@ -180,7 +180,7 @@ class FermiDirac(Dummy):
  
         # Sort them:
         eps_n = num.sort(eps_n)
-        n = self.ne * nu
+        n = int(self.ne * nu)
         self.epsF = 0.5 * (eps_n[n // 2] + eps_n[(n - 1) // 2])
